Bootstrap from compiler pass instead of Drupal service.
Since Drupal can only be bootstrapped once and it's almost entirely
procedural, a Drupal service class doesn't make much sense.  We're now
bootstrapping Drupal from a CompilerPass, very similar to the way that
the SymfonyExtension works.

The CompilerPass method of bootstrapping also gives us the ability to
bundle features with a Drupal module, and specify which module to use
from the config file.